Refactor one (#3)
* refactor + removed code tat did not make sense

* close to release

* i was bored

* lol

* more refactor

* build fix

+++ Overhaul/Core/ModelTracker.cs
@@ -1,5 +1,4 @@
-﻿using System.Data.SqlClient;
-using System.Reflection;
+﻿using System.Reflection;
 using System.Runtime.CompilerServices;
 using Dapper.Contrib.Extensions;
 using Overhaul.Common;
@@ -12,22 +11,21 @@ namespace Overhaul.Core
     public sealed class ModelTracker : IModelTracker
     {
         internal static ModelTrackerOptions Options = new();
-        private readonly string ConnectionString;
         private readonly ISqlGenerator sqlGenerator;
         private readonly ISchemaManager schemaManager;
         private readonly IEnumerable<TableDefinition> databaseDefinitions;
 
         public ModelTracker(string connectionString, ModelTrackerOptions options = null)
         {
+            ConnectionManager.SetConnectionString(connectionString);
+
             Options = options ?? new ModelTrackerOptions();
 
-            sqlGenerator = new SqlGenerator(connectionString);
+            sqlGenerator = new SqlGenerator();
 
-            databaseDefinitions = LoadDatabaseDefinitons();
+            schemaManager = new SchemaManager();
             
-            schemaManager = new SchemaManager(connectionString);
-
-            ConnectionString = connectionString;
+            databaseDefinitions = LoadDatabaseDefinitons();
         }
 
         public void Track(IEnumerable<Type> types)
@@ -36,21 +34,47 @@ namespace Overhaul.Core
             
             if (definitions.Any() || databaseDefinitions.Any())
             {
-                schemaManager.RunSchemaCreate(definitions.Where(i => 
-                !databaseDefinitions.Any(ii => ii.DefType == i.DefType)));
-
-                // DefType are the same but columns don't match up
-                // overridden equals for tableDef
-                schemaManager.RunSchemaUpdate(definitions.Where(i
-                    => databaseDefinitions.Any(ii => ii.DefType == i.DefType && !i.Equals(ii))), databaseDefinitions);
-                
-                schemaManager.RunSchemaDelete(databaseDefinitions.Where(i => !definitions.Any(ii => i.Equals(ii))));
+                CheckForAddedDefinitions(definitions);
+                CheckForUpdatedDefinitions(definitions);
+                CheckForDeletedDefinitions(definitions);
             }
         }
+        private void CheckForDeletedDefinitions(IEnumerable<TableDefinition> definitions)
+        {
+            var deletedDefinitions = databaseDefinitions.Where(i => 
+            !definitions.Any(ii => ii.TableName == i.TableName));
 
+            if (deletedDefinitions.Any())
+            {
+                schemaManager.RunSchemaDelete(databaseDefinitions.Where(i => 
+                !definitions.Any(ii => i.Equals(ii))));
+            }
+        }
+        private void CheckForUpdatedDefinitions(IEnumerable<TableDefinition> definitions)
+        {
+            var updatedDefinitions = definitions.Where(i
+                                => databaseDefinitions.Any(ii => ii.DefType == i.DefType && !i.Equals(ii)));
+
+            if (updatedDefinitions.Any())
+            {
+                // DefType are the same but columns don't match up
+                // overridden equals for tableDef
+                schemaManager.RunSchemaUpdate(updatedDefinitions, databaseDefinitions);
+            }
+        }
+        private void CheckForAddedDefinitions(IEnumerable<TableDefinition> definitions)
+        {
+            var addedDefinitons = definitions.Where(i =>
+                            !databaseDefinitions.Any(ii => ii.TableName == i.TableName));
+
+            if (addedDefinitons.Any())
+            {
+                schemaManager.RunSchemaCreate(addedDefinitons);
+            }
+        }
         public ICrud GetCrudInstance()
         {
-            return new Crud(databaseDefinitions, ConnectionString);
+            return new Crud(databaseDefinitions);
         }
 
         private IEnumerable<TableDefinition> LoadDatabaseDefinitons()
@@ -107,7 +131,12 @@ namespace Overhaul.Core
         // Only needed when debugging, running test
         public static void DeleteTestTables(Type[] tables, string conn = "", bool deleteDef = false)
         {
-            var sqlGenerator = new SqlGenerator(conn);
+            if (!string.IsNullOrEmpty(conn))
+            {
+                ConnectionManager.SetConnectionString(conn);
+            }
+
+            var sqlGenerator = new SqlGenerator();
             var del = tables.ToList();
 
             if (deleteDef)
@@ -120,7 +149,10 @@ namespace Overhaul.Core
 
             foreach (var table in types)
             {
-                sqlGenerator.DeleteTable(table.TableName);
+                if (sqlGenerator.TableExists(table.TableName))
+                {
+                    sqlGenerator.DeleteTable(table.TableName);
+                }
             }
         }
 #endif
