+++ OverhaulTests/Core/SchemaManagerTests.cs
@@ -22,10 +22,15 @@ namespace Overhaul.Core.Tests
         {
             var connection = TestHelper.GetString("devString");
 
-            sqlGenerator = new SqlGenerator(connection);
-            model = new SchemaManager(connection);
+            ConnectionManager.SetConnectionString(connection);
 
-            sqlGenerator.DeleteTable(TableName);
+            sqlGenerator = new SqlGenerator();
+            model = new SchemaManager();
+
+            if (sqlGenerator.TableExists(TableName))
+            {
+                sqlGenerator.DeleteTable(TableName);
+            }
 
             IModelTracker tracker = new ModelTracker(connection);
